requirements.txt
alembic==1.13.1
appdirs==1.4.4
astroid==3.2.2
backports.entry-points-selectable==1.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docstring-to-markdown==0.15
email_validator==2.1.1
Faker==25.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
httplib2==0.22.0
identify==2.6.0
idna==3.7
importlib_metadata==8.0.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauth2client==4.1.3
packaging==24.0
parso==0.8.4
pip-review==1.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
PyDrive==1.3.1
pyflakes==3.2.0
pylint==3.2.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
ruff==0.5.1
setuptools-metadata==0.1.5
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
subprocess.run==0.0.8
toml==0.10.2
tomlkit==0.12.5
tree-sitter-requirements==0.4.0
typing_extensions==4.11.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.3
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
